+++ src/index.ts
@@ -6,3 +6,4 @@
 export * from "./host.js";
 export * from "./server.js";
 export * from "./client.js";
+export * from "./persist.js";